requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
PyYAML==6.0.2
click==8.1.7
pytest==8.3.3
pytest-flask==1.3.0
